Moved lists all pairs in own util module
